# src/api/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Medicamenta.me Public API",
    "version": "1.0.0",
    "description": "RESTful API for healthcare partner integrations. Manage patients, medications, adherence tracking, and more.",
    "contact": {
      "name": "API Support",
      "email": "api-support@medicamenta.me",
      "url": "https://docs.medicamenta.me"
    },
    "license": {
      "name": "Proprietary",
      "url": "https://medicamenta.me/terms"
    }
  },
  "servers": [
    {
      "url": "https://us-central1-{project-id}.cloudfunctions.net/api",
      "description": "Production"
    },
    {
      "url": "http://localhost:5001/{project-id}/us-central1/api",
      "description": "Local Development"
    }
  ],
  "security": [
    {
      "ApiKeyAuth": []
    },
    {
      "BearerAuth": []
    }
  ],
...
  "paths": {
...
    "/v1/patients": {
      "post": {
        "tags": ["Patients"],
        "summary": "Create patient",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatientInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Patient created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Patient"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "get": {
        "tags": ["Patients"],
        "summary": "List patients",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20,
              "maximum": 100
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 0
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["active", "inactive", "deleted"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of patients",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Patient"
                      }
                    },
                    "pagination": {
                      "$ref": "#/components/schemas/Pagination"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/medications": {
      "post": {
        "tags": ["Medications"],
        "summary": "Create medication",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MedicationInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Medication created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Medication"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["Medications"],
        "summary": "List medications",
        "parameters": [
          {
            "name": "patientId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["active", "completed", "deleted"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of medications",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Medication"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/adherence/{patientId}": {
      "get": {
        "tags": ["Adherence"],
        "summary": "Get adherence rate",
        "parameters": [
          {
            "$ref": "#/components/parameters/PatientId"
          },
          {
            "name": "startDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Adherence metrics",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdherenceMetrics"
                }
              }
            }
          }
        }
      }
    },
    "/v1/adherence/confirm": {
      "post": {
        "tags": ["Adherence"],
        "summary": "Confirm dose taken",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["patientId", "medicationId", "scheduledTime"],
                "properties": {
                  "patientId": {
                    "type": "string"
                  },
                  "medicationId": {
                    "type": "string"
                  },
                  "scheduledTime": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "takenAt": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "notes": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Dose confirmed"
          }
        }
      }
    },
    "/v1/webhooks": {
      "post": {
        "tags": ["Webhooks"],
        "summary": "Create webhook",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebhookInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Webhook created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Webhook"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["Webhooks"],
        "summary": "List webhooks",
        "responses": {
          "200": {
            "description": "List of webhooks"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key for authentication"
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token from /v1/auth/token endpoint"
      }
    },
    "parameters": {
      "PatientId": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        },
        "description": "Patient ID"
      }
    },
    "schemas": {
...
      "PatientInput": {
        "type": "object",
        "required": ["name", "dateOfBirth"],
        "properties": {
          "name": {
            "type": "string",
            "example": "João Silva"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "phone": {
            "type": "string"
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date"
          },
          "gender": {
            "type": "string",
            "enum": ["M", "F", "Other"]
          },
          "medicalConditions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "allergies": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "Patient": {
        "allOf": [
          {
            "$ref": "#/components/schemas/PatientInput"
          },
          {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "status": {
                "type": "string",
                "enum": ["active", "inactive", "deleted"]
              },
              "createdAt": {
                "type": "string",
                "format": "date-time"
              },
              "updatedAt": {
                "type": "string",
                "format": "date-time"
              }
            }
          }
        ]
      },
      "MedicationInput": {
        "type": "object",
        "required": ["patientId", "name", "dosage", "frequency"],
        "properties": {
          "patientId": {
            "type": "string"
          },
          "name": {
            "type": "string",
            "example": "Losartana"
          },
          "dosage": {
            "type": "string",
            "example": "50mg"
          },
          "frequency": {
            "type": "string",
            "enum": ["daily", "twice_daily", "three_times_daily", "custom"]
          },
          "times": {
            "type": "array",
            "items": {
              "type": "string",
              "pattern": "^([01]?[0-9]|2[0-3]):[0-5][0-9]$"
            },
            "example": ["08:00", "20:00"]
          },
          "instructions": {
            "type": "string"
          }
        }
      },
      "Medication": {
        "allOf": [
          {
            "$ref": "#/components/schemas/MedicationInput"
          },
          {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "adherenceRate": {
                "type": "number",
                "format": "float"
              },
              "status": {
                "type": "string"
              }
            }
          }
        ]
      },
      "AdherenceMetrics": {
        "type": "object",
        "properties": {
          "patientId": {
            "type": "string"
          },
          "metrics": {
            "type": "object",
            "properties": {
              "totalDoses": {
                "type": "integer"
              },
              "takenDoses": {
                "type": "integer"
              },
              "missedDoses": {
                "type": "integer"
              },
              "adherenceRate": {
                "type": "number",
                "format": "float"
              }
            }
          }
        }
      },
      "WebhookInput": {
        "type": "object",
        "required": ["url", "events"],
        "properties": {
          "url": {
            "type": "string",
            "format": "uri",
            "example": "https://example.com/webhooks/medicamenta"
          },
          "events": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "patient.created",
                "patient.updated",
                "medication.created",
                "dose.taken",
                "dose.missed"
              ]
            }
          },
          "secret": {
            "type": "string",
            "description": "Optional secret for signature verification"
          }
        }
      },
      "Webhook": {
        "allOf": [
          {
            "$ref": "#/components/schemas/WebhookInput"
          },
          {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "status": {
                "type": "string",
                "enum": ["active", "suspended"]
              }
            }
          }
        ]
      },
      "Pagination": {
        "type": "object",
        "properties": {
          "limit": {
            "type": "integer"
          },
          "offset": {
            "type": "integer"
          },
          "total": {
            "type": "integer"
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "details": {
                "type": "object"
              },
              "timestamp": {
                "type": "string",
                "format": "date-time"
              }
            }
          }
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Bad request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Unauthorized",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    }
  }
}